+++ src/routing.rs
@@ -1,0 +1,12 @@
+pub mod basic;
+pub mod login;
+pub mod overlay;
+pub mod sponsors;
+pub mod team;
+pub mod teaminfo;
+pub mod time;
+pub mod visibility;
+pub mod websockets;
+pub mod downs;
+pub mod api;
+pub mod updating;
